Export MtpService for binding from UsbService
Currently MtpService is started by MtpReceiver, but this will be frozen
by activity manager in certain conditions. To prevent this we are going
to bind to MtpService from the UsbService.

Test: Verified the following
 - MTP and PTP works correctly.
 - Process android.process.media is not being frozen.
 - 3P apps without MANAGE_USB permission are not able to start this service.
Bug: 332256525
Flag: NONE - Currently it is not possible to add flags for 2 versions of the service - http://screen/5FXSnVa93Nv47NN.png

Change-Id: I6c9edc391d9c45ee72643b3f753c8d049d1f8566