[MTP] Handling java crash in ServiceIntentSender
MtpDocumentsProvider is in the background when it tries to start a
service, this is not allowed as of Android 12+ and throws a
BackgroundServiceStartNotAllowedException. The short term fix to prevent
the java crash is to handle the exception safely. Since this exception
extends IllegalStateException, catching IllegalStateException is
sufficient.

Bug: 433348162
Flag: EXEMPT BUGFIX
Test: Couldn't repro java crash
Change-Id: I212e6525c7847ec8df4d47fa037598d73711ae68